backend/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
supabase==2.0.2
PyJWT==2.8.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
